# app/api/src/openapi.yaml
openapi: 3.0.3
info:
  title: Identity Atlas Ingest API
  version: 3.4.0
  description: |
    REST API for ingesting authorization data from any source system into Identity Atlas.
    Crawlers authenticate with API keys (Bearer token) and POST batches of records.
    The API handles validation, bulk merge, scoped delete detection, and temporal versioning.

    **Scope:** this specification covers only the **ingest + crawler-admin** surface —
    the endpoints crawlers and the crawler-management UI call. The authenticated
    **read API** consumed by the web UI (matrix & permissions, identities, resources,
    governance & business roles, risk scores, contexts, entity detail, org-chart, tags,
    account-linking — 100+ routes) is **not** described here; see the API reference under
    `docs/api/` for those.
  contact:
    name: Fortigi
    url: https://github.com/Fortigi/IdentityAtlas
servers:
  - url: /api
    description: Relative to deployment root
security:
  - CrawlerApiKey: []
...
paths:
...
  /ingest/systems:
    post:
      tags: [Ingest]
      summary: Ingest systems
      operationId: ingestSystems
      description: Upsert system records. Systems don't require a systemId in the envelope.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          description: Ingest successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /ingest/principals:
    post:
      tags: [Ingest]
      summary: Ingest principals
      operationId: ingestPrincipals
      description: Upsert principal records (users, service principals, etc.). Scope by principalType for full sync deletes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
        '400':
          $ref: '#/components/responses/ValidationError'
  /ingest/resources:
    post:
      tags: [Ingest]
      summary: Ingest resources
      operationId: ingestResources
      description: Upsert resource records (groups, roles, apps, business roles). Scope by resourceType.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/resource-assignments:
    post:
      tags: [Ingest]
      summary: Ingest resource assignments
      operationId: ingestResourceAssignments
      description: Upsert assignments (who has access to what). Composite key (resourceId, principalId, assignmentType). Scope by assignmentType.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/resource-relationships:
    post:
      tags: [Ingest]
      summary: Ingest resource relationships
      operationId: ingestResourceRelationships
      description: Upsert resource-to-resource links (Contains, GrantsAccessTo). Scope by relationshipType.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/identities:
    post:
      tags: [Ingest]
      summary: Ingest identities
      operationId: ingestIdentities
      description: Upsert real-person identity records.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/identity-members:
    post:
      tags: [Ingest]
      summary: Ingest identity members
      operationId: ingestIdentityMembers
      description: Upsert identity-to-principal links.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/contexts:
    post:
      tags: [Ingest]
      summary: Ingest contexts
      operationId: ingestContexts
      description: Upsert organizational contexts (departments, OUs, classifications). Scope by contextType.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/governance/catalogs:
    post:
      tags: [Ingest]
      summary: Ingest governance catalogs
      operationId: ingestGovernanceCatalogs
      description: Upsert governance catalogs (access-package catalogs / IGA sources).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/governance/policies:
    post:
      tags: [Ingest]
      summary: Ingest assignment policies
      operationId: ingestAssignmentPolicies
      description: Upsert assignment policies (approval and settings for business-role assignment).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/governance/requests:
    post:
      tags: [Ingest]
      summary: Ingest assignment requests
      operationId: ingestAssignmentRequests
      description: Upsert assignment requests (access requests and their state).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
  /ingest/governance/certifications:
    post:
      tags: [Ingest]
      summary: Ingest certification decisions
      operationId: ingestCertificationDecisions
      description: Upsert certification decisions (access-review outcomes).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestEnvelope'
      responses:
        '201':
          $ref: '#/components/responses/IngestSuccess'
...
components:
  securitySchemes:
    CrawlerApiKey:
      type: http
      scheme: bearer
      description: "Crawler API key (format: fgc_<random>)"
    EntraIdBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Entra ID JWT token
  schemas:
...
    IngestEnvelope:
      type: object
      required: [records]
      properties:
        systemId:
          type: integer
          description: Target system ID (required for all entity types except systems)
        syncMode:
          type: string
          enum: [full, delta]
          default: delta
          description: "full: merge + delete missing records. delta: merge only."
        scope:
          type: object
          description: 'Narrow delete scope (e.g., { resourceType: "Group" })'
          additionalProperties:
            type: string
        idGeneration:
          type: string
          enum: [native, deterministic]
          default: native
        idPrefix:
          type: string
          description: Namespace for deterministic GUID generation
        syncSession:
          type: string
          enum: [start, continue, end]
          description: For chunked uploads across multiple requests
        syncId:
          type: string
          format: uuid
          description: Session ID (required for continue/end)
        records:
          type: array
          maxItems: 50000
          items:
            type: object
            additionalProperties: true
    IngestResult:
      type: object
      properties:
        table:
          type: string
        inserted:
          type: integer
        updated:
          type: integer
        deleted:
          type: integer
        records:
          type: integer
        durationMs:
          type: integer
...
  responses:
...
    IngestSuccess:
      description: Ingest successful
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/IngestResult'
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              details:
                type: array
                items:
                  type: string
    Unauthorized:
      description: Missing or invalid API key
    Forbidden:
      description: Insufficient permissions or system access
...